+++ product-sdk/packages/contracts/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @parity/product-sdk-contracts
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [4c13257]
+- Updated dependencies [4c13257]
+  - @parity/product-sdk-keys@0.3.0
+  - @parity/product-sdk-signer@0.3.0
+  - @parity/product-sdk-tx@0.2.4
+
 ## 0.5.0
 
 ### Minor Changes